==> 0-1 Scrapy_gameresult_2011.ipynb <==
from selenium import webdriver
import urllib.request as req
from bs4 import BeautifulSoup
import pandas as pd

driver = webdriver.Chrome('./chromedriver')
driver.get('http://www.football-lineups.com/match/74983/')

html = driver.page_source
soup = BeautifulSoup(html,'html.parser')
team1 = soup.select('#mainarea > tbody > tr > td.TDmain > table > tbody > tr > td > font > b' )[0].text
team2 = soup.select('#mainarea > tbody > tr > td.TDmain > table > tbody > tr > td > font > b' )[1].text
print(team1)
print(team2)

html = driver.page_source
soup = BeautifulSoup(html,'html.parser')
time = soup.select('#mainarea > tbody > tr > td.TDmain > table > tbody > tr > td' )[6].text.split()
print(time[0:2])
print(time[1])

html = driver.page_source
soup = BeautifulSoup(html,'html.parser')
weather = soup.select('#mainarea > tbody > tr > td.TDmain > table > tbody > tr > td > font' )[3:]
print(weather[0].text)
print(weather[1].text)

html = driver.page_source
soup = BeautifulSoup(html,'html.parser')
detail = soup.select('#mainarea > tbody > tr > td.TDmain > table >\
                   tbody > tr > td > table > tbody > tr > td > table > tbody > tr > td' )[186:]
print(detail[3].text)
print(detail[4].text)
print(detail[5].text)
print(detail[6].text)
print(detail[7].text)
print(detail[8].text)
print(detail[15].text)
print(detail[16].text)
print(detail[17].text)

74983,75006

driver = webdriver.Chrome('./chromedriver')
team1List = []
team2List = []
timeList = []
weather1List = []
weather2List = []
Shots1List = []
Shots2List = []
Fouls1List = []
Fouls2List = []
possession1List = []
possession2List = []

for i in range(74983,75004):
    url_base = 'http://www.football-lineups.com/match/'
    url_sub = str(i)
    driver.get(url_base + url_sub)
    html = driver.page_source
    soup = BeautifulSoup(html,'html.parser')
    team1 = soup.select('#mainarea > tbody > tr > td.TDmain > table > tbody > tr > td > font > b' )[0].text
    team1List.append(team1)
    team2 = soup.select('#mainarea > tbody > tr > td.TDmain > table > tbody > tr > td > font > b' )[1].text
    team2List.append(team2)
    time = soup.select('#mainarea > tbody > tr > td.TDmain > table > tbody > tr > td' )[6].text.split()
    timeList.append(time[0:2])
    weather = soup.select('#mainarea > tbody > tr > td.TDmain > table > tbody > tr > td > font' )[3:]
    weather1List.append(weather[0].text)
    weather2List.append(weather[1].text)
    detail = soup.select('#mainarea > tbody > tr > td.TDmain > table >\
                   tbody > tr > td > table > tbody > tr > td > table > tbody > tr > td' )[186:]
    Shots1List.append(detail[3].text)
    Shots2List.append(detail[5].text)
    Fouls1List.append(detail[6].text)
    Fouls2List.append(detail[8].text)
    possession1List.append(detail[15].text)
    possession2List.append(detail[17].text)

print(team1List)
print(team2List)
print(timeList)
print(weather1List)
print(weather2List)
print(Shots1List)
print(Shots2List)
print(Fouls1List)
print(Fouls2List)
print(possession1List)
print(possession2List)

len(ShotsList)

gamedetail_2011 = pd.DataFrame(team1List)
gamedetail_2011 = gamedetail_2011.rename(columns = {0:'team1'})
gamedetail_2011['team2'] = team2List
gamedetail_2011['gameday'] = timeList
gamedetail_2011['weather_d'] = weather1List
gamedetail_2011['weather_t'] = weather2List
gamedetail_2011['shoot1'] = Shots1List
gamedetail_2011['shoot2'] = Shots2List
gamedetail_2011['foul1'] = Fouls1List
gamedetail_2011['foul2'] = Fouls2List
gamedetail_2011['possession1'] = possession1List
gamedetail_2011['possession2'] = possession2List

gamedetail_2011

# driver = webdriver.Chrome('./chromedriver')
team1List = []
team2List = []
timeList = []
weather1List = []
weather2List = []
Shots1List = []
Shots2List = []
Fouls1List = []
Fouls2List = []
possession1List = []
possession2List = []

# for i in range(75004):
#     url_base = 'http://www.football-lineups.com/match/'
#     url_sub = str(i)
#     driver.get(url_base + url_sub)
html = driver.page_source
soup = BeautifulSoup(html,'html.parser')
team1 = soup.select('#mainarea > tbody > tr > td.TDmain > table > tbody > tr > td > font > b' )[0].text
team1List.append(team1)
team2 = soup.select('#mainarea > tbody > tr > td.TDmain > table > tbody > tr > td > font > b' )[1].text
team2List.append(team2)
time = soup.select('#mainarea > tbody > tr > td.TDmain > table > tbody > tr > td' )[6].text.split()
timeList.append(time[0:2])
weather = soup.select('#mainarea > tbody > tr > td.TDmain > table > tbody > tr > td > font' )[3:]
weather1List.append(weather[0].text)
#     weather2List.append(weather[1].text)
detail = soup.select('#mainarea > tbody > tr > td.TDmain > table >\
                   tbody > tr > td > table > tbody > tr > td > table > tbody > tr > td' )[186:]
Shots1List.append(detail[3].text)
Shots2List.append(detail[5].text)
Fouls1List.append(detail[6].text)
Fouls2List.append(detail[8].text)
possession1List.append(detail[15].text)
possession2List.append(detail[17].text)

print(team1List)
print(team2List)
print(timeList)
print(weather1List)
print(weather2List)
print(Shots1List)
print(Shots2List)
print(Fouls1List)
print(Fouls2List)
print(possession1List)
print(possession2List)

weather2List = ['0']

gamedetail_aus_2011 = pd.DataFrame(team1List)
gamedetail_aus_2011 = gamedetail_aus_2011.rename(columns = {0:'team1'})
gamedetail_aus_2011['team2'] = team2List
gamedetail_aus_2011['gameday'] = timeList
gamedetail_aus_2011['weather_d'] = weather1List
gamedetail_aus_2011['weather_t'] = weather2List
gamedetail_aus_2011['shoot1'] = Shots1List
gamedetail_aus_2011['shoot2'] = Shots2List
gamedetail_aus_2011['foul1'] = Fouls1List
gamedetail_aus_2011['foul2'] = Fouls2List
gamedetail_aus_2011['possession1'] = possession1List
gamedetail_aus_2011['possession2'] = possession2List

gamedetail_aus_2011

gamedetail_2011_t = pd.concat([gamedetail_2011,gamedetail_aus_2011])

# driver = webdriver.Chrome('./chromedriver')
team1List = []
team2List = []
timeList = []
weather1List = []
weather2List = []
Shots1List = []
Shots2List = []
Fouls1List = []
Fouls2List = []
possession1List = []
possession2List = []

for i in range(75005,75007):
    url_base = 'http://www.football-lineups.com/match/'
    url_sub = str(i)
    driver.get(url_base + url_sub)
    html = driver.page_source
    soup = BeautifulSoup(html,'html.parser')
    team1 = soup.select('#mainarea > tbody > tr > td.TDmain > table > tbody > tr > td > font > b' )[0].text
    team1List.append(team1)
    team2 = soup.select('#mainarea > tbody > tr > td.TDmain > table > tbody > tr > td > font > b' )[1].text
    team2List.append(team2)
    time = soup.select('#mainarea > tbody > tr > td.TDmain > table > tbody > tr > td' )[6].text.split()
    timeList.append(time[0:2])
    weather = soup.select('#mainarea > tbody > tr > td.TDmain > table > tbody > tr > td > font' )[3:]
    weather1List.append(weather[0].text)
    weather2List.append(weather[1].text)
    detail = soup.select('#mainarea > tbody > tr > td.TDmain > table >\
                   tbody > tr > td > table > tbody > tr > td > table > tbody > tr > td' )[186:]
    Shots1List.append(detail[3].text)
    Shots2List.append(detail[5].text)
    Fouls1List.append(detail[6].text)
    Fouls2List.append(detail[8].text)
    possession1List.append(detail[15].text)
    possession2List.append(detail[17].text)

print(team1List)
print(team2List)
print(timeList)
print(weather1List)
print(weather2List)
print(Shots1List)
print(Shots2List)
print(Fouls1List)
print(Fouls2List)
print(possession1List)
print(possession2List)

gamedetail_two_2011 = pd.DataFrame(team1List)
gamedetail_two_2011 = gamedetail_two_2011.rename(columns = {0:'team1'})
gamedetail_two_2011['team2'] = team2List
gamedetail_two_2011['gameday'] = timeList
gamedetail_two_2011['weather_d'] = weather1List
gamedetail_two_2011['weather_t'] = weather2List
gamedetail_two_2011['shoot1'] = Shots1List
gamedetail_two_2011['shoot2'] = Shots2List
gamedetail_two_2011['foul1'] = Fouls1List
gamedetail_two_2011['foul2'] = Fouls2List
gamedetail_two_2011['possession1'] = possession1List
gamedetail_two_2011['possession2'] = possession2List

gamedetail_two_2011

gamedetail_2011_t = pd.concat([gamedetail_2011_t,gamedetail_two_2011])

# gamedetail_2011_t.to_csv('./gamedetail_2011.csv',index=False,header=True)

# driver = webdriver.Chrome('./chromedriver')
team1List = []
team2List = []
timeList = []
weather1List = []
weather2List = []
Shots1List = []
Shots2List = []
Fouls1List = []
Fouls2List = []
possession1List = []
possession2List = []

for i in range(94020,94024):
    url_base = 'http://www.football-lineups.com/match/'
    url_sub = str(i)
    driver.get(url_base + url_sub)
    html = driver.page_source
    soup = BeautifulSoup(html,'html.parser')
    team1 = soup.select('#mainarea > tbody > tr > td.TDmain > table > tbody > tr > td > font > b' )[0].text
    team1List.append(team1)
    team2 = soup.select('#mainarea > tbody > tr > td.TDmain > table > tbody > tr > td > font > b' )[1].text
    team2List.append(team2)
    time = soup.select('#mainarea > tbody > tr > td.TDmain > table > tbody > tr > td' )[6].text.split()
    timeList.append(time[0:2])
    weather = soup.select('#mainarea > tbody > tr > td.TDmain > table > tbody > tr > td > font' )[3:]
    weather1List.append(weather[0].text)
    weather2List.append(weather[1].text)
    detail = soup.select('#mainarea > tbody > tr > td.TDmain > table >\
                   tbody > tr > td > table > tbody > tr > td > table > tbody > tr > td' )[186:]
    Shots1List.append(detail[3].text)
    Shots2List.append(detail[5].text)
    Fouls1List.append(detail[6].text)
    Fouls2List.append(detail[8].text)
    possession1List.append(detail[15].text)
    possession2List.append(detail[17].text)

print(team1List)
print(team2List)
print(timeList)
print(weather1List)
print(weather2List)
print(Shots1List)
print(Shots2List)
print(Fouls1List)
print(Fouls2List)
print(possession1List)
print(possession2List)

gamedetail_qf_2011 = pd.DataFrame(team1List)
gamedetail_qf_2011 = gamedetail_qf_2011.rename(columns = {0:'team1'})
gamedetail_qf_2011['team2'] = team2List
gamedetail_qf_2011['gameday'] = timeList
gamedetail_qf_2011['weather_d'] = weather1List
gamedetail_qf_2011['weather_t'] = weather2List
gamedetail_qf_2011['shoot1'] = Shots1List
gamedetail_qf_2011['shoot2'] = Shots2List
gamedetail_qf_2011['foul1'] = Fouls1List
gamedetail_qf_2011['foul2'] = Fouls2List
gamedetail_qf_2011['possession1'] = possession1List
gamedetail_qf_2011['possession2'] = possession2List

gamedetail_qf_2011

gamedetail_2011_t = pd.concat([gamedetail_2011_t,gamedetail_qf_2011])

# driver = webdriver.Chrome('./chromedriver')
team1List = []
team2List = []
timeList = []
weather1List = []
weather2List = []
Shots1List = []
Shots2List = []
Fouls1List = []
Fouls2List = []
possession1List = []
possession2List = []

for i in range(94338,94340):
    url_base = 'http://www.football-lineups.com/match/'
    url_sub = str(i)
    driver.get(url_base + url_sub)
    html = driver.page_source
    soup = BeautifulSoup(html,'html.parser')
    team1 = soup.select('#mainarea > tbody > tr > td.TDmain > table > tbody > tr > td > font > b' )[0].text
    team1List.append(team1)
    team2 = soup.select('#mainarea > tbody > tr > td.TDmain > table > tbody > tr > td > font > b' )[1].text
    team2List.append(team2)
    time = soup.select('#mainarea > tbody > tr > td.TDmain > table > tbody > tr > td' )[6].text.split()
    timeList.append(time[0:2])
    weather = soup.select('#mainarea > tbody > tr > td.TDmain > table > tbody > tr > td > font' )[3:]
    weather1List.append(weather[0].text)
    weather2List.append(weather[1].text)
    detail = soup.select('#mainarea > tbody > tr > td.TDmain > table >\
                   tbody > tr > td > table > tbody > tr > td > table > tbody > tr > td' )[186:]
    Shots1List.append(detail[3].text)
    Shots2List.append(detail[5].text)
    Fouls1List.append(detail[6].text)
    Fouls2List.append(detail[8].text)
    possession1List.append(detail[15].text)
    possession2List.append(detail[17].text)

print(team1List)
print(team2List)
print(timeList)
print(weather1List)
print(weather2List)
print(Shots1List)
print(Shots2List)
print(Fouls1List)
print(Fouls2List)
print(possession1List)
print(possession2List)

gamedetail_sf_2011 = pd.DataFrame(team1List)
gamedetail_sf_2011 = gamedetail_sf_2011.rename(columns = {0:'team1'})
gamedetail_sf_2011['team2'] = team2List
gamedetail_sf_2011['gameday'] = timeList
gamedetail_sf_2011['weather_d'] = weather1List
gamedetail_sf_2011['weather_t'] = weather2List
gamedetail_sf_2011['shoot1'] = Shots1List
gamedetail_sf_2011['shoot2'] = Shots2List
gamedetail_sf_2011['foul1'] = Fouls1List
gamedetail_sf_2011['foul2'] = Fouls2List
gamedetail_sf_2011['possession1'] = possession1List
gamedetail_sf_2011['possession2'] = possession2List

gamedetail_sf_2011

gamedetail_2011_t = pd.concat([gamedetail_2011_t,gamedetail_sf_2011])

# driver = webdriver.Chrome('./chromedriver')
team1List = []
team2List = []
timeList = []
weather1List = []
weather2List = []
Shots1List = []
Shots2List = []
Fouls1List = []
Fouls2List = []
possession1List = []
possession2List = []

for i in range(94404,94406):
    url_base = 'http://www.football-lineups.com/match/'
    url_sub = str(i)
    driver.get(url_base + url_sub)
    html = driver.page_source
    soup = BeautifulSoup(html,'html.parser')
    team1 = soup.select('#mainarea > tbody > tr > td.TDmain > table > tbody > tr > td > font > b' )[0].text
    team1List.append(team1)
    team2 = soup.select('#mainarea > tbody > tr > td.TDmain > table > tbody > tr > td > font > b' )[1].text
    team2List.append(team2)
    time = soup.select('#mainarea > tbody > tr > td.TDmain > table > tbody > tr > td' )[6].text.split()
    timeList.append(time[0:2])
    weather = soup.select('#mainarea > tbody > tr > td.TDmain > table > tbody > tr > td > font' )[3:]
    weather1List.append(weather[0].text)
    weather2List.append(weather[1].text)
    detail = soup.select('#mainarea > tbody > tr > td.TDmain > table >\
                   tbody > tr > td > table > tbody > tr > td > table > tbody > tr > td' )[186:]
    Shots1List.append(detail[3].text)
    Shots2List.append(detail[5].text)
    Fouls1List.append(detail[6].text)
    Fouls2List.append(detail[8].text)
    possession1List.append(detail[15].text)
    possession2List.append(detail[17].text)

print(team1List)
print(team2List)
print(timeList)
print(weather1List)
print(weather2List)
print(Shots1List)
print(Shots2List)
print(Fouls1List)
print(Fouls2List)
print(possession1List)
print(possession2List)

gamedetail_f_2011 = pd.DataFrame(team1List)
gamedetail_f_2011 = gamedetail_f_2011.rename(columns = {0:'team1'})
gamedetail_f_2011['team2'] = team2List
gamedetail_f_2011['gameday'] = timeList
gamedetail_f_2011['weather_d'] = weather1List
gamedetail_f_2011['weather_t'] = weather2List
gamedetail_f_2011['shoot1'] = Shots1List
gamedetail_f_2011['shoot2'] = Shots2List
gamedetail_f_2011['foul1'] = Fouls1List
gamedetail_f_2011['foul2'] = Fouls2List
gamedetail_f_2011['possession1'] = possession1List
gamedetail_f_2011['possession2'] = possession2List

gamedetail_f_2011

gamedetail_2011_t = pd.concat([gamedetail_2011_t,gamedetail_f_2011])

# gamedetail_2011_t.to_csv('./gamedetail_2011.csv',index=False,header=True)
