==> 0-2 Scrapy_players2011.ipynb <==
from selenium import webdriver
import urllib.request as req
from bs4 import BeautifulSoup
import pandas as pd

driver = webdriver.Chrome('./chromedriver')
driver.get('https://en.wikipedia.org/wiki/2011_AFC_Asian_Cup_squads')

# 네임 추출 테스트
html = driver.page_source
soup = BeautifulSoup(html,'html.parser')
name = soup.select('#mw-content-text > div > table > tbody > tr > th > a')[346]
age = soup.select('#mw-content-text > div > table > tbody > tr > td')[2].text.split()[4].replace(')','')
cap = soup.select('#mw-content-text > div > table > tbody > tr > td')[3].text.strip()
goal = soup.select('#mw-content-text > div > table > tbody > tr > td')[4].text.strip()
print(name)
print(age)
print(cap)
print(goal)
print(club)

# 부상자 때문에 엔트리가 한명 적어진 인도는 따로 Scarpy
nameList = []
ageList = []
capList = []
goalList = []
clubList = []

html = driver.page_source
soup = BeautifulSoup(html,'html.parser')
for i in range(0,368):
    name = soup.select('#mw-content-text > div > table > tbody > tr > th > a')[i].text
    nameList.append(name)
for i in range(0,1840):
    if i % 5 == 2:
        age = soup.select('#mw-content-text > div > table > tbody > tr > td')[i].text.split()[4].replace(')','')
        ageList.append(age)    
    if i % 5 == 3:
        cap = soup.select('#mw-content-text > div > table > tbody > tr > td')[i].text.strip()
        capList.append(cap)
    if i % 5 == 4:
        club = soup.select('#mw-content-text > div > table > tbody > tr > td')[i].text.strip()
        clubList.append(club)

print(nameList)
print(ageList)
print(capList)
print(clubList)

players_2011 = pd.DataFrame(ageList)
players_2011 = players_2011.rename(columns = {0:'age'})
players_2011['cap'] = capList


len(capList)

len(clubList)

players_2011[players_2011['age']=='(aged']

# 나이값 결측치 제거
players_2011.loc[players_2011['age']=='(age','age'] = 26
players_2011.loc[players_2011['age']=='(aged','age'] = 34

players_2011 = players_2011.astype(int)

players_2011['name'] = nameList

players_2011['club'] = clubList

players_2011.tail()

players_2011.to_csv('./players_2011.csv',index=False,header=True)

players_2011 = pd.read_csv('./players_2011.csv')

qatar = players_2011[0:23]
uzbekistan = players_2011[23:46]
kuwait = players_2011[46:69]
china = players_2011[69:92]
japan = players_2011[92:115]
jordan = players_2011[115:138]
saudi = players_2011[138:161]
syria = players_2011[161:184]
south_korea = players_2011[184:207]
Australia = players_2011[207:230]
bahrain = players_2011[230:253]
india = players_2011[253:276]
north_korea = players_2011[276:299]
uae = players_2011[299:322]
iraq = players_2011[322:345]
iran = players_2011[345:]

iran

qatar['nation'] = 'qatar'
uzbekistan['nation'] = 'uzbekistan'
kuwait['nation'] = 'kuwait'
china['nation'] = 'china'
japan['nation'] = 'japan'
jordan['nation'] = 'jordan'
saudi['nation'] = 'saudi'
syria['nation'] = 'syria'
south_korea['nation'] = 'south_korea'
Australia['nation'] = 'Australia'
bahrain['nation'] = 'bahrain'
india['nation'] = 'india'
north_korea['nation'] = 'north_korea'
uae['nation'] = 'uae'
iraq['nation'] = 'iraq'
iran['nation'] = 'iran'

players_2011 = pd.concat([qatar,uzbekistan,kuwait,china,japan,jordan,saudi,syria,south_korea,Australia,bahrain,india
                     ,north_korea,uae,iraq,iran])

colnames = ['name','age','cap','club','nation']
players_2011[colnames]

korea_players['nation'] = 'south-korea'

korea_players.info()

colnames = ['name','age','height','cap','goal','salary','nation']
korea_players[colnames]

players_2011.to_csv('./semi_players_2011.csv',index=False,header=True)

# database로 넣기
import pymysql
from sqlalchemy import create_engine

# Table의 column이름과 df의 column 이름이 일치해야함.
column = colnames = ['age','cap','goal','height','name','nation','salary']
total_players = pd.read_csv('./total_players.csv')
total_players = pd.DataFrame(total_players)
total_players.columns = column
total_players.head()


engine = create_engine("mysql+pymysql://root:" + "1111" + 
                       "@127.0.0.1:3306/asiancup?charset=utf8",
                      encoding = 'utf-8')
conn = engine.connect()


total_players.to_sql(name = 'players', con = engine, if_exists = 'append', index = False)


conn.close()
