==> 0-2 Scrapy_players2011_02.ipynb <==
from selenium import webdriver
import urllib.request as req
from bs4 import BeautifulSoup
import pandas as pd

driver = webdriver.Chrome('./chromedriver')
driver.get('https://www.transfermarkt.co.uk/usbekistan/startseite/verein/3563/saison_id/2010')

# 네임 추출 테스트
html = driver.page_source
soup = BeautifulSoup(html,'html.parser')
backnum = soup.select('#yw1 > table > tbody > tr > td ')[0].text.strip()
name = soup.select('#yw1 > table > tbody > tr > td ')[1].text.strip()
age = soup.select('#yw1 > table > tbody > tr > td')[29].text.split()[0].replace('(','').replace(')','')
height = soup.select('#yw1 > table > tbody > tr > td ')[4].text.strip().replace(',','').replace('m','')
cap = soup.select('#yw1 > table > tbody > tr > td ')[6].text.strip()
goal = soup.select('#yw1 > table > tbody > tr > td ')[7].text.strip()
salary = soup.select('#yw1 > table > tbody > tr > td ')[9].text.strip().replace('£','').replace('k','000')
print(backnum)
print(name)
print(age)
print(height)
print(cap)
print(goal)
print(salary)


# 부상자 때문에 엔트리가 두명 적어진 한국은 따로 Scarpy
backnumList = []
nameList = []
ageList = []
heightList = []
capList = []
goalList = []
salaryList = []

html = driver.page_source
soup = BeautifulSoup(html,'html.parser')
for i in range(0,330):
    if i % 10 == 1:
        name = soup.select('#yw1 > table > tbody > tr > td')[i].text.split()
        nameList.append(name)
    if i % 10 == 2:
        age = soup.select('#yw1 > table > tbody > tr > td')[i].text.split()[3].replace('(','').replace(')','')
        ageList.append(age)    
    if i % 10 == 4:
        height = soup.select('#yw1 > table > tbody > tr > td')[i].text.replace(',','').replace('m','').strip()
        heightList.append(height)
    if i % 10 == 6:
        cap = soup.select('#yw1 > table > tbody > tr > td')[i].text.replace('-','0').strip()
        capList.append(cap)
    if i % 10 == 7:
        goal = soup.select('#yw1 > table > tbody > tr > td')[i].text.replace('-','0').strip()
        goalList.append(goal)
    if i % 10 == 9:
        salary = soup.select('#yw1 > table > tbody > tr > td')[i].text.replace('.','').replace('£','').replace('k','').replace('m','000').strip()
        salaryList.append(salary)

print(nameList)
print(ageList)
print(heightList)
print(capList)
print(goalList)
print(salaryList)

# 불필요한 이름 정제
# namelist = []
# for i in range(0,21):
#     namelist.append(nameList[i][:2])
# namelist2 = []
# for i in range(0,21):
#     namelist2.append(namelist[i][0]+namelist[i][1])
namelist = []
namelist2 = []
for i in range(0,33):
    namelist.append(nameList[i][:2])
    namelist2.append(namelist[i][0]+namelist[i][1])

players = pd.DataFrame(heightList)
players = players.rename(columns = {0:'height'})
players['age'] = ageList
players['cap'] = capList
players['goal'] = goalList
players['salary'] = salaryList

# players['age'] = players['age'].str[-3:-1]

players.loc[players['height']==''] = 0
players.loc[players['height']=='N/A','height'] = 0
players.loc[players['age'].isnull(),'age'] = 0
players.loc[players['salary']=='-\xa0'] = 0
players.loc[players['salary']=='-','salary'] = 0

players

players = players.astype(int)

# 키 정보가 없는 선수들의 키를 평균값으로 대체
players.loc[players['height']==0,'height'] =int(players['height'][players['height']!=0].mean())

# salary 정보가 없는 선수들의 연봉을 평균값으로 대체
players.loc[players['salary']==0,'salary'] =int(players['salary'][players['salary']!=0].median())

# 나이 정보가 없는 선수들의 키를 평균값으로 대체
players.loc[players['age']==0,'age'] =int(players['age'][players['age']!=0].mean())

players['name'] = namelist2
players['nation'] = 'uae'

players.head()

colnames = ['name','age','height','cap','goal','salary','nation']
players[colnames]

# players.to_csv('./uzbekistan_2011.csv',index=False,header=True)
# players.to_csv('./qatar_2011.csv',index=False,header=True)
# players.to_csv('./china_2011.csv',index=False,header=True)
# players.to_csv('./kuwait_2011.csv',index=False,header=True)
# players.to_csv('./uae_2011.csv',index=False,header=True)

players = pd.concat([nation2[0],nation2[1],nation2[2],india_players,nation2[3],nation2[4],nation2[5],nation2[6],
                     korea_players,nation2[7],nation2[8],nation2[9],nation2[10],nation2[11],nation2[12],
                     nation2[13],nation2[14],nation2[15],nation2[16],nation2[17],nation2[18],nation2[19],
                     nation2[20],nation2[21]])

colnames = ['name','age','height','cap','goal','salary','nation']
players[colnames]

# players.to_csv('./total_players.csv',index=False,header=True)

saudi_2011 = pd.read_csv('./saudi_2011.csv')
australia_2011 = pd.read_csv('./australia_2011.csv')
barain_2011 = pd.read_csv('./Bahrain_2011.csv')
china_2011 = pd.read_csv('./china_2011.csv')
india_2011 = pd.read_csv('./india_2011.csv')
irak_2011 = pd.read_csv('./irak_2011.csv')
iran_2011 = pd.read_csv('./iran_2011.csv')
japan_2011 = pd.read_csv('./japan_2011.csv')
jordan_2011 = pd.read_csv('./jordan_2011.csv')
kuwait_2011 = pd.read_csv('./kuwait_2011.csv')
north_2011 = pd.read_csv('./nordkorea_2011.csv')
quatar_2011 = pd.read_csv('./qatar_2011.csv')
syria_2011 = pd.read_csv('./syria_2011.csv')
south_2011 = pd.read_csv('./sudkorea_2011.csv')
uae_2011 = pd.read_csv('./uae_2011.csv')
uz_2011 = pd.read_csv('./uzbekistan_2011.csv')

players_2011 = pd.concat([quatar_2011,uz_2011,kuwait_2011,china_2011,japan_2011,jordan_2011,saudi_2011,syria_2011,
                          south_2011,australia_2011,barain_2011,india_2011,north_2011,uae_2011,irak_2011,iran_2011])

del players_2011['age']
del players_2011['cap']
del players_2011['goal']
del players_2011['nation']

players_2011_main = pd.read_csv('./semi_players_2011.csv')

len(players_2011)

len(players_2011_main)

players_2011_total = pd.merge(players_2011_main, players_2011, on = 'name')

# players_2011_total.to_csv('./total_players_2011.csv',index=False,header=True)

players_2011_total.head()

# database로 넣기
import pymysql
from sqlalchemy import create_engine

# Table의 column이름과 df의 column 이름이 일치해야함.
column = colnames = ['age','cap','name','club','nation','height','salary']
total_players = pd.read_csv('./total_players_2011.csv')
total_players = pd.DataFrame(total_players)
total_players.columns = column
total_players.head()


engine = create_engine("mysql+pymysql://root:" + "1111" + 
                       "@127.0.0.1:3306/asiancup?charset=utf8",
                      encoding = 'utf-8')
conn = engine.connect()


# total_players.to_sql(name = 'players_2011', con = engine, if_exists = 'append', index = False)


# conn.close()
