==> 0-2 Scrapy_players2015(ㅇㅕㄴㅂㅗㅇㅍㅗㅎㅏㅁ)_merge_Joo.ipynb <==
from selenium import webdriver
import urllib.request as req
from bs4 import BeautifulSoup
import pandas as pd

korea_china_2015 = pd.read_csv('./korea_china_players_2015.csv')
uzbek_2015 = pd.read_csv('./Data/uzbek_player_2015.csv')
australia_2015 = pd.read_csv('./Data/australia_player_2015.csv')
oman_irak_2015 = pd.read_csv('./Data/oman_irak_player_2015.csv')
pala_katar_2015 = pd.read_csv('./Data/pala_katar_player_2015.csv')
kuwait_2015 = pd.read_csv('./Data/kuwait_player_2015.csv')
saudi_2015 = pd.read_csv('./Data/saudi_player_2015.csv')
north_korea_2015 = pd.read_csv('./Data/north-korea_player_2015.csv')
iran_2015 = pd.read_csv('./Data/iran_player_2015.csv')
uae_2015 = pd.read_csv('./Data/uae_player_2015.csv')
bahrain_2015 = pd.read_csv('./Data/bahrain_player_2015.csv')
japan_2015 = pd.read_csv('./Data/japan_player_2015.csv')
jordan_2015 = pd.read_csv('./Data/jordan_player_2015.csv')

players_2015 = pd.concat([korea_china_2015,uzbek_2015,australia_2015,oman_irak_2015,pala_katar_2015,kuwait_2015,
                         saudi_2015,north_korea_2015,iran_2015,uae_2015,bahrain_2015,japan_2015,jordan_2015])

del players_2015['age']
del players_2015['cap']
del players_2015['goal']
del players_2015['nation']

players_2015_main = pd.read_csv('./Data/semi_players_2015.csv')

len(players_2015)

len(players_2015_main)

players_2015_total = pd.merge(players_2015_main, players_2015, on = 'name')

players_2015_total.to_csv('./Data/total_players_2015.csv',index=False,header=True)

players_2015_total

# database로 넣기
import pymysql
from sqlalchemy import create_engine

# Table의 column이름과 df의 column 이름이 일치해야함.
column = colnames = ['age','cap','goal','name','club','nation','height','salary']
total_players = pd.read_csv('./Data/total_players_2015.csv')
total_players = pd.DataFrame(total_players)
total_players.columns = column
total_players.head()

engine = create_engine("mysql+pymysql://root:" + "1111" + 
                       "@192.168.35.143:3306/asiancup?charset=utf8",
                      encoding = 'utf-8')
conn = engine.connect()


total_players.to_sql(name = 'players_2015', con = engine, if_exists = 'append', index = False)
