==> 0-2 Scrapy_players2015.ipynb <==
from selenium import webdriver
import urllib.request as req
from bs4 import BeautifulSoup
import pandas as pd

driver = webdriver.Chrome('./chromedriver')
driver.get('https://en.wikipedia.org/wiki/2015_AFC_Asian_Cup_squads')

# 네임 추출 테스트
html = driver.page_source
soup = BeautifulSoup(html,'html.parser')
name = soup.select('#mw-content-text > div > table > tbody > tr > th > a')[22].text
age = soup.select('#mw-content-text > div > table > tbody > tr > td')[2].text.split()[4].replace(')','')
cap = soup.select('#mw-content-text > div > table > tbody > tr > td')[3].text.split()[0]
goal = soup.select('#mw-content-text > div > table > tbody > tr > td')[4].text
club = soup.select('#mw-content-text > div > table > tbody > tr > td')[5].text
print(name)
print(age)
print(cap)
print(goal)
print(club)

# 부상자 때문에 엔트리가 한명 적어진 인도는 따로 Scarpy
nameList = []
ageList = []
capList = []
goalList = []
clubList = []

html = driver.page_source
soup = BeautifulSoup(html,'html.parser')
for i in range(0,367):
    name = soup.select('#mw-content-text > div > table > tbody > tr > th > a')[i].text
    nameList.append(name)
for i in range(0,2202):
    if i % 6 == 2:
        age = soup.select('#mw-content-text > div > table > tbody > tr > td')[i].text.split()[4].replace(')','')
        ageList.append(age)    
    if i % 6 == 3:
        cap = soup.select('#mw-content-text > div > table > tbody > tr > td')[i].text.strip()
        capList.append(cap)
    if i % 6 == 4:
        goal = soup.select('#mw-content-text > div > table > tbody > tr > td')[i].text.strip()
        goalList.append(goal)
    if i % 6 == 5:
        club = soup.select('#mw-content-text > div > table > tbody > tr > td')[i].text.strip()
        clubList.append(club)

print(nameList)
print(ageList)
print(capList)
print(goalList)
print(clubList)

players_2015 = pd.DataFrame(ageList)
players_2015 = players_2015.rename(columns = {0:'age'})
players_2015['cap'] = capList
players_2015['goal'] = goalList

# goal 결측치 제거
players_2015[players_2015['goal']=='']

players_2015.loc[players_2015['goal']=='','goal'] = 0

players_2015 = players_2015.astype(int)

players_2015['name'] = nameList

players_2015['club'] = clubList

players_2015.head()

# players_2015.to_csv('./players_2015.csv',index=False,header=True)

players_2015 = pd.read_csv('./players_2015.csv')

Australia = players_2015[0:23]
south_korea = players_2015[23:46]
oman = players_2015[46:69]
kuwait = players_2015[69:92]
uzbekistan = players_2015[92:115]
saudi = players_2015[115:138]
china = players_2015[138:161]
north_korea = players_2015[161:183]
iran = players_2015[183:206]
uae = players_2015[206:229]
qatar = players_2015[229:252]

bahrain = players_2015[252:275]
japan = players_2015[275:298]
jordan = players_2015[298:321]
iraq = players_2015[321:344]
palestine = players_2015[344:]

palestine

Australia['nation'] = 'Australia'
south_korea['nation'] = 'south_korea'
oman['nation'] = 'oman'
kuwait['nation'] = 'kuwait'
uzbekistan['nation'] = 'uzbekistan'
saudi['nation'] = 'saudi'
china['nation'] = 'china'
north_korea['nation'] = 'north_korea'
iran['nation'] = 'iran'
uae['nation'] = 'uae'
qatar['nation'] = 'qatar'
bahrain['nation'] = 'bahrain'
japan['nation'] = 'japan'
jordan['nation'] = 'jordan'
iraq['nation'] = 'iraq'
palestine['nation'] = 'palestine'



players_2015 = pd.concat([Australia,south_korea,oman,kuwait,uzbekistan,saudi,china,north_korea,iran,uae,qatar,bahrain
                     ,japan,jordan,iraq,palestine])

colnames = ['name','age','cap','goal','club','nation']
players_2015[colnames]

players_2015.to_csv('./semi_players_2015.csv',index=False,header=True)

# database로 넣기
import pymysql
from sqlalchemy import create_engine

# Table의 column이름과 df의 column 이름이 일치해야함.
column = colnames = ['age','cap','goal','height','name','nation','salary']
total_players = pd.read_csv('./total_players.csv')
total_players = pd.DataFrame(total_players)
total_players.columns = column
total_players.head()


engine = create_engine("mysql+pymysql://root:" + "1111" + 
                       "@127.0.0.1:3306/asiancup?charset=utf8",
                      encoding = 'utf-8')
conn = engine.connect()


total_players.to_sql(name = 'players', con = engine, if_exists = 'append', index = False)


conn.close()
