==> 0-2 Scrapy_players2019.ipynb <==
from selenium import webdriver
import urllib.request as req
from bs4 import BeautifulSoup
import pandas as pd

driver = webdriver.Chrome('./chromedriver')
driver.get('https://www.transfermarkt.co.uk/south-korea/kader/verein/3589/saison_id/2018/plus/1')

html = driver.page_source
soup = BeautifulSoup(html,'html.parser')
team = soup.select('#yw1 > table > tbody > tr > td' )[3]
print(team)

team

# 네임 추출 테스트
html = driver.page_source
soup = BeautifulSoup(html,'html.parser')
name = soup.select('#yw1 > table > tbody > tr > td ')[201].text.strip()
age = soup.select('#yw1 > table > tbody > tr > td')[202].text.split()[3].replace('(','').replace(')','')
height = soup.select('#yw1 > table > tbody > tr > td ')[204].text.strip().replace(',','').replace('m','')
cap = soup.select('#yw1 > table > tbody > tr > td ')[206].text.strip()
goal = soup.select('#yw1 > table > tbody > tr > td ')[207].text.strip()
salary = soup.select('#yw1 > table > tbody > tr > td ')[209].text.strip().replace('£','').replace('k','000')
print(name)
print(age)
print(height)
print(cap)
print(goal)
print(salary)


# 부상자 때문에 엔트리가 두명 적어진 한국은 따로 Scarpy
nameList = []
ageList = []
heightList = []
capList = []
goalList = []
salaryList = []

html = driver.page_source
soup = BeautifulSoup(html,'html.parser')
for i in range(0,210):
    if i % 10 == 1:
        name = soup.select('#yw1 > table > tbody > tr > td')[i].text.split()
        nameList.append(name)
    if i % 10 == 2:
        age = soup.select('#yw1 > table > tbody > tr > td')[i].text.split()[3].replace('(','').replace(')','')
        ageList.append(age)    
    if i % 10 == 4:
        height = soup.select('#yw1 > table > tbody > tr > td')[i].text.replace(',','').replace('m','')
        heightList.append(height)
    if i % 10 == 6:
        cap = soup.select('#yw1 > table > tbody > tr > td')[i].text
        capList.append(cap)
    if i % 10 == 7:
        goal = soup.select('#yw1 > table > tbody > tr > td')[i].text.replace('-','0')
        goalList.append(goal)
    if i % 10 == 9:
        salary = soup.select('#yw1 > table > tbody > tr > td')[i].text.replace('.','').replace('£','').replace('k','').replace('m','000')
        salaryList.append(salary)

print(nameList)
print(ageList)
print(heightList)
print(capList)
print(goalList)
print(salaryList)

# 불필요한 이름 정제
# namelist = []
# for i in range(0,21):
#     namelist.append(nameList[i][:2])
# namelist2 = []
# for i in range(0,21):
#     namelist2.append(namelist[i][0]+namelist[i][1])
namelist = []
namelist2 = []
for i in range(0,21):
    namelist.append(nameList[i][:2])
    namelist2.append(namelist[i][0]+namelist[i][1])

namelist2

korea_players = pd.DataFrame(heightList)
korea_players['age'] = ageList
korea_players['cap'] = capList
korea_players['goal'] = goalList
korea_players['salary'] = salaryList
korea_players = korea_players.astype(int)

korea_players['name'] = namelist2
korea_players = korea_players.rename(columns = {0:'height'})


korea_players['nation'] = 'south-korea'

korea_players.info()

colnames = ['name','age','height','cap','goal','salary','nation']
korea_players[colnames]

korea_players.to_csv('./korea_players.csv',index=False,header=True)

driver = webdriver.Chrome('./chromedriver')
driver.get('https://www.transfermarkt.co.uk/indien/kader/verein/13957/saison_id/2018/plus/1')

# 부상자 때문에 엔트리가 한명 적어진 인도는 따로 Scarpy
nameList = []
ageList = []
heightList = []
capList = []
goalList = []
salaryList = []

html = driver.page_source
soup = BeautifulSoup(html,'html.parser')
for i in range(0,220):
    if i % 10 == 1:
        name = soup.select('#yw1 > table > tbody > tr > td')[i].text.split()
        nameList.append(name)
    if i % 10 == 2:
        age = soup.select('#yw1 > table > tbody > tr > td')[i].text.split()[3].replace('(','').replace(')','')
        ageList.append(age)    
    if i % 10 == 4:
        height = soup.select('#yw1 > table > tbody > tr > td')[i].text.replace(',','').replace('m','').replace(' ','-')
        heightList.append(height)
    if i % 10 == 6:
        cap = soup.select('#yw1 > table > tbody > tr > td')[i].text.replace('-','0')
        capList.append(cap)
    if i % 10 == 7:
        goal = soup.select('#yw1 > table > tbody > tr > td')[i].text.replace('-','0')
        goalList.append(goal)
    if i % 10 == 9:
        salary = soup.select('#yw1 > table > tbody > tr > td')[i].text.replace('.','').replace('£','').replace('k','').replace('m','000').replace('-','0')
        salaryList.append(salary)

print(nameList)
print(ageList)
print(heightList)
print(capList)
print(goalList)
print(salaryList)

namelist = []
namelist2 = []
for i in range(0,22):
    namelist.append(nameList[i][:2])
    namelist2.append(namelist[i][0]+namelist[i][1])

india_players = pd.DataFrame(heightList)
india_players = india_players.rename(columns = {0:'height'})
india_players['age'] = ageList
india_players['cap'] = capList
india_players['goal'] = goalList
india_players['salary'] = salaryList

india_players['height'] = india_players['height'].str[0:3]

india_players.loc[india_players['height']=='-','height'] = 0

india_players = india_players.astype(int)

india_players.info()

# 키 정보가 없는 선수들의 키를 평균값으로 대체
india_players.loc[india_players['height']==0,'height'] =int(india_players['height'][india_players['height']!=0].mean())

india_players['name'] = namelist2

# salary 정보가 없는 선수들의 연봉을 평균값으로 대체
india_players.loc[india_players['salary']==0,'salary'] =int(india_players['salary'][india_players['salary']!=0].mean())

india_players['nation'] = 'india'

colnames = ['name','age','height','cap','goal','salary','nation']
india_players[colnames]

india_players.to_csv('./india_players.csv',index=False,header=True)

# driver = webdriver.Chrome('./chromedriver')

nation = ['vereinigte-arabische-emirate', 'thailand', 'bahrain', 'jordanien', 'australien', 'palastina',
         'syrien', 'china', 'kirgisistan', 'philippinen', 'iran', 'irak', 'vietnam', 'jemen', 'katar',
         'saudi-arabien', 'libanon', 'nordkorea', 'japan', 'usbekistan', 'oman', 'turkmenistan']
nationid = ['5147', '5676', '7214', '15737', '3433', '17758','13674', '5598', '3956', '15234', '3582', 
            '3560', '8778', '15922', '14162', '3807', '3586', '15457', '3435', '3563', '14165', '14248']
nameList = []
namelist = []
nationList = []
ageList = []
heightList = []
capList = []
goalList = []
salaryList = []

for i in range(len(nation)):
    print(nation[i], nationid[i])
    url_base = 'https://www.transfermarkt.co.uk/'
    url_sub = nation[i] + '/kader/verein/' + nationid[i] + '/saison_id/2018/plus/1'
    driver.get(url_base + url_sub)
    html = driver.page_source
    soup = BeautifulSoup(html,'html.parser')
    for j in range(0,23):
        nationList.append(nation[i])
    for i in range(0,231):
        if i % 10 == 1:
            name = soup.select('#yw1 > table > tbody > tr > td')[i].text.replace('-',' ').split()
            nameList.append(name)
        if i % 10 == 2:
            age = soup.select('#yw1 > table > tbody > tr > td')[i].text.split()[3].replace('(','').replace(')','')
            ageList.append(age)    
        if i % 10 == 4:
            height = soup.select('#yw1 > table > tbody > tr > td')[i].text.replace(',','').replace('m','').replace(' ','-')
            heightList.append(height)
        if i % 10 == 6:
            cap = soup.select('#yw1 > table > tbody > tr > td')[i].text.replace('-','0')
            capList.append(cap)
        if i % 10 == 7:
            goal = soup.select('#yw1 > table > tbody > tr > td')[i].text.replace('-','0')
            goalList.append(goal)
        if i % 10 == 9:
            salary = soup.select('#yw1 > table > tbody > tr > td')[i].text.replace('.','').replace('£','').replace('k','').replace('m','000').replace('-','0')
            salaryList.append(salary)

nameList[66] = ['Yusuf','Yusuf']

namelist = []
namelist2 = []
for i in range(0,506):
    namelist.append(nameList[i][:2])
    namelist2.append(namelist[i][0]+namelist[i][1])

# dataframe으로 변환
player = pd.DataFrame(
        {
            'age':ageList,
            'height':heightList,
            'cap':capList,
            'goal':goalList,
            'salary':salaryList,
        }
)
player.head()

# 키가 1.9m로 되어있는 한선수 변경
player.loc[player['height']=='19-','height'] = '190'

player['height'] = player['height'].str[0:3]

player.loc[player['height']=='-','height'] = 0

player = player.astype(int)

player['name'] = namelist2
player['nation'] = nationList

player.loc[player["nation"] == 'vereinigte-arabische-emirate','nation'] = "U.A.E"
player.loc[player["nation"] == 'jordanien','nation'] = "jordan"
player.loc[player["nation"] == 'australien','nation'] = "australia"
player.loc[player["nation"] == 'palastina','nation'] = "palastine"
player.loc[player["nation"] == 'syrien','nation'] = "syria"
player.loc[player["nation"] == 'kirgisistan','nation'] = "kyrgyzstan"
player.loc[player["nation"] == 'philippinen','nation'] = "philippines"
player.loc[player["nation"] == 'irak','nation'] = "iraq"
player.loc[player["nation"] == 'katar','nation'] = "Qatar"
player.loc[player["nation"] == 'saudi-arabien','nation'] = "saudi-arabia"
player.loc[player["nation"] == 'libanon','nation'] = "lebanon"
player.loc[player["nation"] == 'nordkorea','nation'] = "north-korea"
player.loc[player["nation"] == 'usbekistan','nation'] = "uzbekistan"

colnames = ['name','age','height','cap','goal','salary','nation']
player[colnames]

# 각 나라별 결측치 값 정리

nation2 = player['nation'].unique()
nation2

for i in range(len(nation2)):
    nation2[i] = player[player['nation']==nation2[i]]

for i in range(len(nation2)):
# 키 정보가 없는 선수들의 키를 평균값으로 대체
    nation2[i].loc[nation2[i]['height']==0,'height'] =int(nation2[i]['height'][nation2[i]['height']!=0].mean())
    # salary 정보가 없는 선수들의 연봉을 평균값으로 대체
    nation2[i].loc[nation2[i]['salary']==0,'salary'] =int(nation2[i]['salary'][nation2[i]['salary']!=0].mean())

players = pd.concat([nation2[0],nation2[1],nation2[2],india_players,nation2[3],nation2[4],nation2[5],nation2[6],
                     korea_players,nation2[7],nation2[8],nation2[9],nation2[10],nation2[11],nation2[12],
                     nation2[13],nation2[14],nation2[15],nation2[16],nation2[17],nation2[18],nation2[19],
                     nation2[20],nation2[21]])

colnames = ['name','age','height','cap','goal','salary','nation']
players[colnames]

players.to_csv('./total_players.csv',index=False,header=True)

# database로 넣기
import pymysql
from sqlalchemy import create_engine

# Table의 column이름과 df의 column 이름이 일치해야함.
column = colnames = ['age','cap','goal','height','name','nation','salary']
total_players = pd.read_csv('./total_players.csv')
total_players = pd.DataFrame(total_players)
total_players.columns = column
total_players.head()


engine = create_engine("mysql+pymysql://root:" + "1111" + 
                       "@127.0.0.1:3306/asiancup?charset=utf8",
                      encoding = 'utf-8')
conn = engine.connect()


total_players.to_sql(name = 'players_2019', con = engine, if_exists = 'append', index = False)


conn.close()
