==> 1 TeamData_chaging_for_machine_ver1 - choi.ipynb <==
%matplotlib inline
import numpy as np
import pandas as pd
import matplotlib
import matplotlib.pyplot as plt

# 선수데이터
a=pd.read_csv('./total_players.csv')
a.head()

# 아시안컵 19년도
asian = pd.read_csv('./data/AsianCup19.csv')
asian.head()

len(asian)

asian['팀'].unique()

# asian 테이블에 국가 뒤에 공백있는거 제거
names = []
for i in asian['팀']:
    names.append(i.strip())

print(names)

asian['팀'] = names

# 피파랭킹 19년도
fifarank = pd.read_csv('./data/FIFARank.csv')
fifarank = fifarank.drop(columns = ['점수'])
fifarank.head()

fifarank['국가'].unique()

# 한국이라 되어있기 때문에 아시안컵 테이블에 있는 대한민국으로 바꿔주기
fifarank.loc[fifarank['국가'] == '한국', '국가'] = '대한민국'

# 피파랭킹 추가하기
tot_ver1 = pd.merge(asian, fifarank, left_on='팀', right_on='국가', how='left')
tot_ver1.head()

# null값 확인
tot_ver1.isnull().sum()

# 연봉 합계 추가하기
money = a.pivot_table('salary',index='nation', aggfunc=np.sum)
money.head()

money = money.reset_index()
money.head()

money['nation']

tot_ver1['팀'].unique()

# 수작업..
money.loc[money['nation'] == 'Qatar', 'nation'] = '카타르'
money.loc[money['nation'] == 'U.A.E', 'nation'] = '아랍에미리트'
money.loc[money['nation'] == 'australia', 'nation'] = '호주'
money.loc[money['nation'] == 'bahrain', 'nation'] = '바레인'
money.loc[money['nation'] == 'china', 'nation'] = '중국'
money.loc[money['nation'] == 'india', 'nation'] = '인도'

money.loc[money['nation'] == 'iran', 'nation'] = '이란'
money.loc[money['nation'] == 'iraq', 'nation'] = '이라크'
money.loc[money['nation'] == 'japan', 'nation'] = '일본'
money.loc[money['nation'] == 'jemen', 'nation'] = '예멘'

money.loc[money['nation'] == 'jordan', 'nation'] = '요르단'
money.loc[money['nation'] == 'kyrgyzstan', 'nation'] = '키르기스스탄'
money.loc[money['nation'] == 'lebanon', 'nation'] = '레바논'
money.loc[money['nation'] == 'north-korea', 'nation'] = '북한'

money.loc[money['nation'] == 'oman', 'nation'] = '오만'
money.loc[money['nation'] == 'palastine', 'nation'] = '팔레스타인'
money.loc[money['nation'] == 'philippines', 'nation'] = '필리핀'

money.loc[money['nation'] == 'saudi-arabia', 'nation'] = '사우디아라비아'
money.loc[money['nation'] == 'south-korea', 'nation'] = '대한민국'
money.loc[money['nation'] == 'syria', 'nation'] = '시리아'

money.loc[money['nation'] == 'thailand', 'nation'] = '태국'
money.loc[money['nation'] == 'turkmenistan', 'nation'] = '투르크메니스탄'
money.loc[money['nation'] == 'uzbekistan', 'nation'] = '우즈베키스탄'
money.loc[money['nation'] == 'vietnam', 'nation'] = '베트남'

# 연봉 단위는 '천원'임 (0 세개 뺀 값 -> 너무 단위가 커서)
money.head()

# 나라이름 중복되는 컬럼 지우기
tot_ver1 = tot_ver1.drop(columns=['국가'])
tot_ver1.head()

# 합치기
tot_ver2 = pd.merge(tot_ver1,money,left_on='팀', right_on='nation', how='left')
tot_ver2 = tot_ver2.drop(columns=['nation'])
tot_ver2.head()

pd.set_option('display.float_format', '{:.5f}'.format) # 지수표현 없이 보는 것
tot_ver2.describe()

tot_ver2 = tot_ver2.drop(columns=['팀1점수'])
tot_ver2 = tot_ver2.drop(columns=['팀2점수'])
tot_ver2.head()

tot_ver2['salary'].describe()

tot_ver2['salary'].value_counts()

tot_ver2['랭킹'].unique()

tot_ver2['랭킹']

# 카테고리로 변경
tot_ver2['랭킹'] = tot_ver2['랭킹'].astype('category')

tot_ver2.info()

# 원래 랭킹이 낮은 것 부터 1 ~ 23까지
tot_ver2['랭킹'].cat.categories = ['%s'%i for i in range(1,24)]

tot_ver2.head()

tot_ver2['salary'].unique()

## * 각 칼럼의 값이 낮을 수록 팀에 유리하다고 가정하고 데이터 변형 -
==> 랭킹 1위는 숫자상으로 가장 낮은 값이지만 가장 유리하다. -> 모든칼럼적용

# 승패 데이터 변경
