==> 4 machine-learning(2019)_final.ipynb <==
# Module
import pandas as pd
import numpy as np
from scipy import stats # 확률분포 그래프
from sklearn.ensemble import RandomForestRegressor
from sklearn.tree import DecisionTreeClassifier
from sklearn.model_selection import train_test_split
from sklearn import metrics

import warnings
warnings.filterwarnings('ignore')

total = pd.read_csv('./data/total_match.csv')
total_av = pd.read_csv('./data/total_av.csv')

# 카테고리 순서
natcat = pd.DataFrame(list(total['team'].unique()))
natcat = natcat.sort_values(by = 0)
# natcat.reindex(index = range(1,19), columns=[0])
natcat['순서'] = range(1,19)
natcat

total.head()

# 국가를 카테고리화 시키기
total['team']=total['team'].astype('category')
total['weater']=total['weater'].astype('category')
total['team'].cat.categories = ['%s'%i for i in range(1,19)]

total.info()

gameday	team	shoot	foul	possession	weater	temp	rank	salary	goal

target_col = ['team','shoot','foul','possession','weater','temp','rank','salary']
total[target_col].head()

train, test = train_test_split(total,test_size=0.3,random_state=0)

# Train data 구성
# 모델 제작용 컬럼과 정답 컬럼을 구분
train_X = train[target_col]
train_Y = train['goal']
train_Y = train_Y.astype(float)
train_X.head()

train_Y.head()

X_test = test[target_col]
Y_test = test['goal']
Y_test = Y_test.astype(float)
X_test.head()

Y_test.head()

feature_one = train_X.values
target = train_Y.values

# model 생성
model = RandomForestRegressor(n_estimators=100, n_jobs=-1,random_state=0)
model

# Training
model.fit(feature_one, target)

# 예측
pre = model.predict(X_test)
pre

13	3	14	55	4	20	17	4180995
team	shoot	foul	possession	weater	temp	rank	salary

# # 2015년 호주, 한국 예측(1:2)
# pre = model.predict([[7,4,26,53,3,22,20,1534325]])
# pre2 = model.predict([[18,3,19,47,3,22,31,2308955]])
# print('2015년 호주, 한국 예측')
# print(pre)
# print(pre2)
# # 2015년 호주, 쿠웨이트 예측(4:1)
# pre = model.predict([[7,8,19,60,3,16,100,1534325]])
# pre2 = model.predict([[16,5,10,40,3,16,139,5148]])
# print('2015년 호주, 쿠웨이트 예측')
# print(pre)
# print(pre2)
# # 2015년 한국 오만 예측(1:0)
# pre = model.predict([[18,6,14,67,3,19,51,2306476]])
# pre2 = model.predict([[6,1,8,33,3,19,104,165857]])
# print('2015년 한국 오만 예측')
# print(pre)
# print(pre2)
# # 2019년 이란 일본 예측(0:3)
# pre = model.predict([[11,2,17,46,1,31,45,4530110]])
# pre2 = model.predict([[13,4,18,54,1,31,53,6934420]])
# print('2019년 이란 일본 예측')
# print(pre)
# print(pre2)
# # 2019년 카타르 일본 예측(3:1)
# pre = model.predict([[15,3,8,40,1,31,93,147021]])
# pre2 = model.predict([[13,1,18,60,1,31,50,6934420]])
# print('2019년 카타르 일본 예측(')
# print(pre)
# print(pre2)
# # 2019년 한국 카타르 예측(0:1)
# pre = model.predict([[18,2,19,58,1,31,53,6038313]])
# pre2 = model.predict([[15,4,14,42,1,31,93,147021]])
# print('2019년 한국 카타르 예측')
# print(pre)
# print(pre2)
# # 2019년 이란 중국 예측(3:0)
# pre = model.predict([[11,7,19,55,1,31,29,4530110]])
# pre2 = model.predict([[14,1,15,45,1,31,76,326027]])
# print('2019년 이란 중국 예측')
# print(pre)
# print(pre2)

total_av['salary'].max()

# 예측
# 날씨 - 구름 1 안개 2 비 3 맑음 4
team_list = list(total_av['team'])

print('바레인', '북한', '사우디아라비아', '시리아', '아랍에미리트', '오만', '오스트레일리아', '요르단')
print('우즈베키스탄', '이라크', '이란', '인도', '일본', '중국', '카타르', '쿠웨이트', '팔레스타인','한국')
print('')

print('')
print('오늘의 날씨를 입력해주세요!!','구름', '안개', '비', '맑음')

weather1 = input("날씨(구름', '안개', '비', '맑음) : ")
if weather1 == '구름':
    weather1 = 1
if weather1 == '안개':
    weather1 = 2
if weather1 == '비':
    weather1 = 3
if weather1 == '맑음':
    weather1 = 4
temp1 = input("오늘의 기온 : ")

print('')

# 팀 입력받기
team1 = input("팀 1 국가 : ")
a = natcat.loc[natcat[0]=='%s'%team1,"순서"]
a=int(a)
for i in team_list:
    if team1 == i:
        shoot1 = total_av['shoot'][total_av['team']==i]
        foul1 = total_av['foul'][total_av['team']==i]
        possession1 = total_av['possession'][total_av['team']==i]
        rank1 = total_av['rank'][total_av['team']==i]
        salary1 = total_av['salary'][total_av['team']==i]

print('')

# 팀2
team2 = input("팀 2 국가 : ")
b = natcat.loc[natcat[0]=='%s'%team2,"순서"]
b=int(b)
for i in team_list:
    if team2 == i:
        shoot2 = total_av['shoot'][total_av['team']==i]
        foul2 = total_av['foul'][total_av['team']==i]
        possession2 = total_av['possession'][total_av['team']==i]
        rank2 = total_av['rank'][total_av['team']==i]
        salary2 = total_av['salary'][total_av['team']==i]


pre1 = model.predict([[int(a),(int(shoot1)),(int(foul1)),(int(possession1)),int(weather1),int(temp1),(211 - int(rank1)),int(salary1)]])
pre2 = model.predict([[int(b),(int(shoot2)),(int(foul2)),(int(possession2)),int(weather1),int(temp1),(211 - int(rank2)),int(salary2)]])
print('')
print(team1, pre1) # 쿠웨이트
print(team2, pre2) # 우즈벡

if pre1 > pre2:
    print('%s  이 %s 을(를) 이겼습니다.'%(team1,team2))
if pre2 > pre1:
    print('%s  이 %s 을(를) 이겼습니다.'%(team2,team1))
